feat(dashboard): implement dashboard store with fetching capabilities and tests
